+++ src/extension/paging.js
@@ -10,9 +10,9 @@
 
 var IASPagingExtension = function() {
   this.ias = null;
-  this.currentScrollOffset = 0;
   this.pagebreaks = [[0, document.location.toString()]];
   this.lastPageNum = 1;
+  this.enabled = true;
   this.listeners = {
     pageChange: new IASCallbacks()
   };
@@ -24,16 +24,19 @@ var IASPagingExtension = function() {
    * @param scrollThreshold
    */
   this.onScroll = function(currentScrollOffset, scrollThreshold) {
-    var currentPageNum = this.getCurrentPageNum(currentScrollOffset),
+    if (!this.enabled) {
+      return;
+    }
+
+    var ias = this.ias,
+        currentPageNum = this.getCurrentPageNum(currentScrollOffset),
         currentPagebreak = this.getCurrentPagebreak(currentScrollOffset),
         urlPage;
-
-    this.currentScrollOffset = currentScrollOffset;
 
     if (this.lastPageNum !== currentPageNum) {
       urlPage = currentPagebreak[1];
 
-      this.ias.fire('pageChange', [currentPageNum, currentScrollOffset, urlPage]);
+      ias.fire('pageChange', [currentPageNum, currentScrollOffset, urlPage]);
     }
 
     this.lastPageNum = currentPageNum;
@@ -45,14 +48,49 @@ var IASPagingExtension = function() {
    * @param url
    */
   this.onNext = function(url) {
-    this.pagebreaks.push([this.currentScrollOffset, url]);
+    var currentScrollOffset = this.ias.getCurrentScrollOffset(this.ias.$scrollContainer);
+
+    this.pagebreaks.push([currentScrollOffset, url]);
 
     // trigger pageChange and update lastPageNum
-    var currentPageNum = this.getCurrentPageNum(this.currentScrollOffset) + 1;
+    var currentPageNum = this.getCurrentPageNum(currentScrollOffset) + 1;
 
-    this.ias.fire('pageChange', [currentPageNum, this.currentScrollOffset, url]);
+    this.ias.fire('pageChange', [currentPageNum, currentScrollOffset, url]);
 
     this.lastPageNum = currentPageNum;
+  };
+
+  /**
+   * Keeps track of pagebreaks
+   *
+   * @param url
+   */
+  this.onPrev = function(url) {
+    var self = this,
+        ias = self.ias,
+        currentScrollOffset = ias.getCurrentScrollOffset(ias.$scrollContainer),
+        prevCurrentScrollOffset = currentScrollOffset - ias.$scrollContainer.height(),
+        $firstItem = ias.getFirstItem();
+
+    this.enabled = false;
+
+    this.pagebreaks.unshift([0, url]);
+
+    ias.one('rendered', function() {
+      // update pagebreaks
+      for (var i = 1, l = self.pagebreaks.length; i < l; i++) {
+        self.pagebreaks[i][0] = self.pagebreaks[i][0] + $firstItem.offset().top;
+      }
+
+      // trigger pageChange and update lastPageNum
+      var currentPageNum = self.getCurrentPageNum(prevCurrentScrollOffset) + 1;
+
+      ias.fire('pageChange', [currentPageNum, prevCurrentScrollOffset, url]);
+
+      self.lastPageNum = currentPageNum;
+
+      self.enabled = true;
+    });
   };
 
   return this;
@@ -61,14 +99,23 @@ var IASPagingExtension = function() {
 /**
  * @public
  */
-IASPagingExtension.prototype.bind = function(ias) {
+IASPagingExtension.prototype.initialize = function(ias) {
   this.ias = ias;
-
-  ias.on('next', $.proxy(this.onNext, this));
-  ias.on('scroll', $.proxy(this.onScroll, this));
 
   // expose the extensions listeners
   jQuery.extend(ias.listeners, this.listeners);
+};
+
+/**
+ * @public
+ */
+IASPagingExtension.prototype.bind = function(ias) {
+  try {
+    ias.on('prev', $.proxy(this.onPrev, this), this.priority);
+  } catch (exception) {}
+
+  ias.on('next', $.proxy(this.onNext, this), this.priority);
+  ias.on('scroll', $.proxy(this.onScroll, this), this.priority);
 };
 
 /**
@@ -93,8 +140,7 @@ IASPagingExtension.prototype.getCurrentPageNum = function(scrollOffset) {
  * @param {number} scrollOffset
  * @returns {number}|null
  */
-IASPagingExtension.prototype.getCurrentPagebreak = function(scrollOffset)
-{
+IASPagingExtension.prototype.getCurrentPagebreak = function(scrollOffset) {
   for (var i = (this.pagebreaks.length - 1); i >= 0; i--) {
     if (scrollOffset > this.pagebreaks[i][0]) {
       return this.pagebreaks[i];
@@ -103,3 +149,9 @@ IASPagingExtension.prototype.getCurrentPagebreak = function(scrollOffset)
 
   return null;
 };
+
+/**
+ * @public
+ * @type {number}
+ */
+IASPagingExtension.prototype.priority = 500;