+++ test/03-extensions-test.js
@@ -18,13 +18,16 @@ describe("IAS", function () {
     jQuery.ias('destroy');
   });
 
-  it("should call bind on extension when loading extension", function() {
+  it("should call initialize on extension when loading extension", function() {
     var deferred = when.defer();
     var spy1 = this.spy();
 
     var anExtension = function() {};
 
     anExtension.prototype.bind = function(ias) {
+      // an extension needs a bind method
+    };
+    anExtension.prototype.initialize = function(ias) {
       spy1.call();
     };
 
@@ -37,6 +40,25 @@ describe("IAS", function () {
     return deferred.promise;
   });
 
+  it("should call bind on extension when initializing", function() {
+    var deferred = when.defer();
+    var spy1 = this.spy();
+
+    var anExtension = function() {};
+
+    anExtension.prototype.bind = function(ias) {
+      spy1.call();
+    };
+
+    jQuery.ias().extension(new anExtension());
+    jQuery.ias().initialize();
+
+    expect(spy1).toHaveBeenCalledOnce();
+
+    deferred.resolve();
+
+    return deferred.promise;
+  });
 
   it("extension can add listeners", function() {
     var anExtension = function() {
@@ -59,6 +81,7 @@ describe("IAS", function () {
 
     // now let's register the extension
     jQuery.ias().extension(new anExtension());
+    jQuery.ias().initialize();
 
     // this should now be possible and not throw an error
     jQuery.ias().on('test', function() {});