+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+## 2.x
+
+* Added History extension
+* Added `ready` event
+* Added `loaded` event (`load` is now triggered before loading starts)
+* Added `rendered` event (`render` is now triggered before rendering starts)
+* Added priority to callbacks
+* Added `initialize` call for extensions
+* Added `one` method
+
 ## 2.0.0
 
 * Completely rewritten